backend/mediateque/requirements.txt
flask==3.0.0
flask-PyMongo==2.3.0
python-dotenv==1.0.0
flask-cors==4.0.0
pytest-flask